refactor(fastpitch): registry-safe torch-free engine module
Move the LightningModule, dataset and collate into
phoonnx_train/fastpitch/lightning.py and defer all torch imports in
phoonnx_train/engines/fastpitch.py until a model is built, so the engine
registry stays importable without the training stack installed.

- extract torch-free pitch (F0) corpus statistics into
  phoonnx_train/fastpitch/pitch_stats.py; malformed or zero-std cached
  stats are now recomputed instead of trusted
- extract torch-free resolve_module_kwargs (quality preset expansion,
  variant defaulting, config precedence) for direct unit testing
- accept a preprocessed dataset directory (not just dataset.jsonl) like
  the other engines; pass explicit batch_size to Lightning log_dict so
  logging works with the custom Batch type
- export the multi-speaker 'speaker' input as int32 to match the
  FastPitchAdapter/MixerTTSAdapter feed dtype; reuse
  torch_compat.onnx_export_kwargs for the exporter selection
- rewrite tests/test_fastpitch_training.py torch-free (registry, preset
  and config handling, pitch-stats edge cases via file-path import)
