+++ phoonnx_train/engines/fastpitch.py
@@ -13,35 +13,27 @@ Durations are learned unsupervised (no external aligner needed) via the
 vendored ``AlignmentNetwork`` + monotonic alignment search, same recipe as
 upstream FastPitch/coqui.
 
-Reuses the shared VITS preprocessing pipeline: ``audio_norm_path`` /
-``audio_spec_path`` (linear spectrogram) are produced by
-``phoonnx_train.preprocess``; the mel target is derived from the linear
-spectrogram at train time via ``phoonnx_train.vits.mel_processing`` (no
-separate mel cache needed). Pitch (F0) is extracted via
-``extra_preprocess`` (pyworld), following the same contract used by the
-OptiSpeech training engine.
+The LightningModule / dataset / collate live in
+``phoonnx_train.fastpitch.lightning``; heavy torch imports are deferred
+until a model is actually built so the engine registry stays importable in
+torch-free environments.
 
 ONNX export follows the contract consumed by ``phoonnx.engines.fastpitch``
 (``FastPitchAdapter``, which reuses ``MixerTTSAdapter``'s feed/parse logic):
-inputs ``token_ids`` [B,T] (+ optional ``speaker`` [B] for multi-speaker) ->
-output ``mel_spec`` [B, 80, T_mel]. This mirrors
+inputs ``token_ids`` [B,T] + ``pace``/``pitch_mul``/``pitch_add`` [1]
+(+ optional ``speaker`` [B] for multi-speaker) -> output ``mel_spec``
+[B, 80, T_mel]. This mirrors
 ``scripts/conversion/coqui_fastpitch_export/export_fp.py``, which exports
 pretrained Coqui FastPitch/SpeedySpeech checkpoints with the same I/O.
 """
 import logging
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple
-
-import pytorch_lightning as pl
-import torch
-from torch import LongTensor
-from torch.utils.data import DataLoader, Dataset, random_split
+from typing import TYPE_CHECKING, Any, Dict, List
 
 from phoonnx_train.engines.base import BaseTrainingEngine, TrainingEngineConfig
-from phoonnx_train.fastpitch.losses import ForwardTTSLoss
-from phoonnx_train.fastpitch.model import ForwardTTS, ForwardTTSArgs
-from phoonnx_train.vits.dataset import PhoonnxDataset, Utterance
-from phoonnx_train.vits.mel_processing import spec_to_mel_torch
+
+if TYPE_CHECKING:  # heavy import — only needed for type annotations
+    import pytorch_lightning as pl
 
 _LOG = logging.getLogger(__name__)
 
@@ -50,7 +42,7 @@ OPSET_VERSION = 14
 
 # Quality tier -> ForwardTTS hyper-param overrides.
 # "variant" selects encoder/decoder family + pitch predictor (set via
-# ``config.extra['variant']``, default "fastpitch").
+# ``config.extra['variant']``, default per registered engine name).
 _QUALITY_PRESETS: Dict[str, Dict[str, Any]] = {
     "x-low": {
         "hidden_channels": 128,
@@ -75,334 +67,38 @@ _QUALITY_PRESETS: Dict[str, Dict[str, Any]] = {
     },
 }
 
-# SpeedySpeech-specific overrides layered on top of the tier preset when
-# ``variant == "speedyspeech"``.
-_SPEEDYSPEECH_OVERRIDES: Dict[str, Any] = {
-    "encoder_type": "residual_conv_bn",
-    "decoder_type": "residual_conv_bn",
-    "use_pitch": False,
-}
 
+def resolve_module_kwargs(
+    config: TrainingEngineConfig,
+    default_variant: str,
+    **kwargs: Any,
+) -> Dict[str, Any]:
+    """Merge shared config + ``extra`` bag + call-site kwargs into the
+    keyword arguments for ``ForwardTTSModule``.
 
-class UtteranceTensors:
-    """Per-utterance training tensors: ids, mel (via spec), optional f0."""
-
-    __slots__ = ("phoneme_ids", "spectrogram", "speaker_id", "pitch")
-
-    def __init__(self, phoneme_ids: LongTensor, spectrogram: torch.Tensor,
-                 speaker_id: Optional[LongTensor], pitch: Optional[torch.Tensor]):
-        self.phoneme_ids = phoneme_ids
-        self.spectrogram = spectrogram
-        self.speaker_id = speaker_id
-        self.pitch = pitch
-
-
-class Batch:
-    __slots__ = (
-        "phoneme_ids", "phoneme_lengths", "mels", "mel_lengths",
-        "pitch", "speaker_ids",
+    Torch-free (pure dict handling) so it is unit-testable without the
+    training stack. ``extra['quality']``, if present, expands to the
+    matching preset (unknown names fall back to ``medium``); explicit
+    ``extra``/kwargs keys win over preset values, and ``variant`` defaults
+    to *default_variant* when not set explicitly.
+    """
+    extra = dict(config.extra)
+    preset_name = extra.pop("quality", None)
+    merged: Dict[str, Any] = {}
+    if preset_name is not None:
+        if preset_name not in _QUALITY_PRESETS:
+            _LOG.warning("unknown quality %r — falling back to 'medium'", preset_name)
+            preset_name = "medium"
+        merged.update(_QUALITY_PRESETS[preset_name])
+    merged.update(extra)
+    merged.update(kwargs)
+    merged.setdefault("variant", default_variant)
+    merged.update(
+        num_symbols=config.num_symbols,
+        num_speakers=config.num_speakers,
+        sample_rate=config.sample_rate,
     )
-
-    def __init__(self, phoneme_ids, phoneme_lengths, mels, mel_lengths, pitch, speaker_ids):
-        self.phoneme_ids = phoneme_ids
-        self.phoneme_lengths = phoneme_lengths
-        self.mels = mels
-        self.mel_lengths = mel_lengths
-        self.pitch = pitch
-        self.speaker_ids = speaker_ids
-
-
-class ForwardTTSDataset(Dataset):
-    """
-    Wraps :class:`phoonnx_train.vits.dataset.PhoonnxDataset` utterances,
-    converting the shared linear-spectrogram cache to mel at load time and
-    optionally loading a pitch (F0) cache produced by
-    :meth:`ForwardTTSTrainingEngine.extra_preprocess`.
-    """
-
-    def __init__(self, dataset_paths: List[Path], mel_channels: int,
-                 filter_length: int, sample_rate: int,
-                 mel_fmin: float, mel_fmax: Optional[float],
-                 max_phoneme_ids: Optional[int] = None):
-        self._inner = PhoonnxDataset(dataset_paths, max_phoneme_ids=max_phoneme_ids)
-        self.mel_channels = mel_channels
-        self.filter_length = filter_length
-        self.sample_rate = sample_rate
-        self.mel_fmin = mel_fmin
-        self.mel_fmax = mel_fmax
-        # corpus F0 statistics over voiced frames — the pitch target is
-        # z-scored so its scale matches the other loss terms and the
-        # pitch_mul/pitch_add inference controls operate on a normalized
-        # quantity (raw Hz would dwarf every other loss)
-        self.pitch_mean, self.pitch_std = self._pitch_stats(dataset_paths)
-
-    def _f0_candidate(self, utt: "Utterance") -> Path:
-        return Path(str(utt.audio_spec_path)).with_suffix("").with_suffix(".f0.npy")
-
-    def _pitch_stats(self, dataset_paths: List[Path]):
-        import json
-
-        import numpy as np
-
-        stats_path = None
-        for p in dataset_paths:
-            p = Path(p)
-            if p.is_dir():
-                stats_path = p / "pitch_stats.json"
-                break
-        if stats_path and stats_path.is_file():
-            stats = json.loads(stats_path.read_text())
-            return float(stats["mean"]), float(stats["std"])
-
-        voiced = []
-        for utt in self._inner.utterances:
-            cand = self._f0_candidate(utt)
-            if cand.exists():
-                f0 = np.load(cand)
-                voiced.append(f0[f0 > 0])
-        if not voiced:
-            return 0.0, 1.0  # no pitch caches — identity normalization
-        allv = np.concatenate(voiced)
-        mean = float(allv.mean()) if allv.size else 0.0
-        std = float(allv.std()) if allv.size else 1.0
-        std = std or 1.0
-        if stats_path:
-            try:
-                stats_path.write_text(json.dumps({"mean": mean, "std": std}))
-            except OSError:
-                pass
-        return mean, std
-
-    def __len__(self) -> int:
-        return len(self._inner)
-
-    def __getitem__(self, idx: int) -> UtteranceTensors:
-        utt: Utterance = self._inner.utterances[idx]
-        spec = torch.load(utt.audio_spec_path)  # [n_fft//2+1, T]
-        mel = spec_to_mel_torch(
-            spec.unsqueeze(0), self.filter_length, self.mel_channels,
-            self.sample_rate, self.mel_fmin, self.mel_fmax,
-        ).squeeze(0)  # [mel_channels, T]
-
-        pitch = None
-        # optional sidecar cache written by extra_preprocess: "<stem>.f0.npy"
-        # next to audio_norm_path's parent cache dir.
-        f0_candidate = self._f0_candidate(utt)
-        if f0_candidate.exists():
-            import numpy as np
-
-            f0 = np.load(f0_candidate).astype("float32")
-            # z-score voiced frames with the corpus stats; unvoiced stays 0
-            voiced = f0 > 0
-            f0[voiced] = (f0[voiced] - self.pitch_mean) / self.pitch_std
-            pitch = torch.from_numpy(f0).unsqueeze(0)  # [1, T_f0]
-
-        return UtteranceTensors(
-            phoneme_ids=LongTensor(utt.phoneme_ids),
-            spectrogram=mel,
-            speaker_id=LongTensor([utt.speaker_id]) if utt.speaker_id is not None else None,
-            pitch=pitch,
-        )
-
-
-class ForwardTTSCollate:
-    def __init__(self, is_multispeaker: bool, use_pitch: bool):
-        self.is_multispeaker = is_multispeaker
-        self.use_pitch = use_pitch
-
-    def __call__(self, utterances: List[UtteranceTensors]) -> Batch:
-        n = len(utterances)
-        max_ph = max(u.phoneme_ids.size(0) for u in utterances)
-        max_mel = max(u.spectrogram.size(1) for u in utterances)
-        mel_channels = utterances[0].spectrogram.size(0)
-
-        phoneme_ids = torch.zeros(n, max_ph, dtype=torch.long)
-        phoneme_lengths = torch.zeros(n, dtype=torch.long)
-        mels = torch.zeros(n, mel_channels, max_mel)
-        mel_lengths = torch.zeros(n, dtype=torch.long)
-        pitch = torch.zeros(n, 1, max_mel) if self.use_pitch else None
-        speaker_ids = torch.zeros(n, dtype=torch.long) if self.is_multispeaker else None
-
-        for i, utt in enumerate(utterances):
-            pl_ = utt.phoneme_ids.size(0)
-            ml = utt.spectrogram.size(1)
-            phoneme_ids[i, :pl_] = utt.phoneme_ids
-            phoneme_lengths[i] = pl_
-            mels[i, :, :ml] = utt.spectrogram
-            mel_lengths[i] = ml
-            if self.use_pitch and utt.pitch is not None:
-                pl_len = min(utt.pitch.size(-1), max_mel)
-                pitch[i, :, :pl_len] = utt.pitch[:, :pl_len]
-            if speaker_ids is not None and utt.speaker_id is not None:
-                speaker_ids[i] = utt.speaker_id
-
-        return Batch(phoneme_ids, phoneme_lengths, mels, mel_lengths, pitch, speaker_ids)
-
-
-class ForwardTTSModule(pl.LightningModule):
-    """LightningModule wrapping the vendored :class:`ForwardTTS` model."""
-
-    def __init__(
-        self,
-        num_symbols: int,
-        num_speakers: int = 1,
-        sample_rate: int = 22050,
-        filter_length: int = 1024,
-        mel_channels: int = 80,
-        mel_fmin: float = 0.0,
-        mel_fmax: Optional[float] = None,
-        variant: str = "fastpitch",
-        hidden_channels: int = 384,
-        hidden_channels_ffn: int = 1024,
-        encoder_num_layers: int = 6,
-        decoder_num_layers: int = 6,
-        num_heads: int = 1,
-        use_pitch: Optional[bool] = None,
-        use_energy: bool = False,
-        dataset: Optional[List[Path]] = None,
-        learning_rate: float = 1e-4,
-        betas: Tuple[float, float] = (0.9, 0.998),
-        eps: float = 1e-9,
-        weight_decay: float = 1e-6,
-        batch_size: int = 8,
-        num_workers: int = 1,
-        validation_split: float = 0.1,
-        num_test_examples: int = 5,
-        max_phoneme_ids: Optional[int] = None,
-        binary_loss_start_epoch: int = 0,
-        binary_loss_warmup_epochs: int = 10,
-        **kwargs: Any,
-    ):
-        super().__init__()
-        self.save_hyperparameters()
-
-        variant_overrides: Dict[str, Any] = {}
-        if variant == "speedyspeech":
-            variant_overrides.update(_SPEEDYSPEECH_OVERRIDES)
-        if use_pitch is not None:
-            variant_overrides["use_pitch"] = use_pitch
-
-        args = ForwardTTSArgs(
-            num_chars=num_symbols,
-            out_channels=mel_channels,
-            hidden_channels=hidden_channels,
-            hidden_channels_ffn=hidden_channels_ffn,
-            encoder_num_layers=encoder_num_layers,
-            decoder_num_layers=decoder_num_layers,
-            num_heads=num_heads,
-            use_energy=use_energy,
-            num_speakers=num_speakers,
-            **variant_overrides,
-        )
-        self.model_args = args
-        self.model = ForwardTTS(args)
-        self.criterion = ForwardTTSLoss()
-
-        self._train_dataset: Optional[Dataset] = None
-        self._val_dataset: Optional[Dataset] = None
-        self._test_dataset: Optional[Dataset] = None
-        self._load_datasets(validation_split, num_test_examples, max_phoneme_ids)
-
-    # ------------------------------------------------------------------
-    def _load_datasets(self, validation_split: float, num_test_examples: int,
-                       max_phoneme_ids: Optional[int]) -> None:
-        if not self.hparams.dataset:
-            _LOG.debug("No dataset to load")
-            return
-        full_dataset = ForwardTTSDataset(
-            self.hparams.dataset,
-            mel_channels=self.hparams.mel_channels,
-            filter_length=self.hparams.filter_length,
-            sample_rate=self.hparams.sample_rate,
-            mel_fmin=self.hparams.mel_fmin,
-            mel_fmax=self.hparams.mel_fmax,
-            max_phoneme_ids=max_phoneme_ids,
-        )
-        valid_size = max(0, int(len(full_dataset) * validation_split))
-        test_size = min(num_test_examples, max(0, len(full_dataset) - valid_size))
-        train_size = len(full_dataset) - valid_size - test_size
-        self._train_dataset, self._test_dataset, self._val_dataset = random_split(
-            full_dataset, [train_size, test_size, valid_size]
-        )
-
-    def _collate(self) -> ForwardTTSCollate:
-        return ForwardTTSCollate(
-            is_multispeaker=self.hparams.num_speakers > 1,
-            use_pitch=bool(self.model_args.use_pitch),
-        )
-
-    def train_dataloader(self):
-        return DataLoader(self._train_dataset, collate_fn=self._collate(),
-                          num_workers=self.hparams.num_workers,
-                          batch_size=self.hparams.batch_size, shuffle=True)
-
-    def val_dataloader(self):
-        return DataLoader(self._val_dataset, collate_fn=self._collate(),
-                          num_workers=self.hparams.num_workers,
-                          batch_size=self.hparams.batch_size)
-
-    def test_dataloader(self):
-        return DataLoader(self._test_dataset, collate_fn=self._collate(),
-                          num_workers=self.hparams.num_workers,
-                          batch_size=self.hparams.batch_size)
-
-    # ------------------------------------------------------------------
-    def forward(self, x, pace: float = 1.0, pitch_mul: float = 1.0,
-               pitch_add: float = 0.0, speaker=None):
-        return self.model.inference(x, speaker=speaker, pace=pace,
-                                    pitch_mul=pitch_mul, pitch_add=pitch_add)
-
-    def _step(self, batch: Batch, binary_loss_weight: float = 1.0) -> Dict[str, torch.Tensor]:
-        outputs = self.model(
-            x=batch.phoneme_ids,
-            x_lengths=batch.phoneme_lengths,
-            y=batch.mels,
-            y_lengths=batch.mel_lengths,
-            pitch=batch.pitch,
-            speaker=batch.speaker_ids,
-        )
-        losses = self.criterion(
-            decoder_output=outputs["model_outputs"],
-            decoder_target=batch.mels.transpose(1, 2),
-            decoder_output_lens=batch.mel_lengths,
-            dur_output=outputs["durations_log"],
-            dur_target=outputs["durations"],
-            input_lens=batch.phoneme_lengths,
-            pitch_output=outputs["pitch_avg_pred"],
-            pitch_target=outputs["pitch_avg"],
-            aligner_logprob=outputs["alignment_logprob"],
-            alignment_hard=outputs["alignment_hard"],
-            alignment_soft=outputs["alignment_soft"],
-            binary_loss_weight=binary_loss_weight,
-        )
-        return losses
-
-    def training_step(self, batch: Batch, batch_idx: int):
-        # ramp the binary alignment loss in over binary_loss_warmup_epochs
-        # starting at binary_loss_start_epoch — full-strength binarization
-        # against a still-random soft alignment destabilizes the aligner
-        warmup = max(1, self.hparams.binary_loss_warmup_epochs)
-        binary_loss_weight = min(1.0, max(
-            0.0, (self.current_epoch - self.hparams.binary_loss_start_epoch + 1) / warmup))
-        losses = self._step(batch, binary_loss_weight=binary_loss_weight)
-        self.log_dict({f"train_{k}": v for k, v in losses.items()}, prog_bar=False)
-        return losses["loss"]
-
-    def validation_step(self, batch: Batch, batch_idx: int):
-        losses = self._step(batch)
-        self.log_dict({f"val_{k}": v for k, v in losses.items()})
-        return losses["loss"]
-
-    def configure_optimizers(self):
-        optimizer = torch.optim.AdamW(
-            self.model.parameters(),
-            lr=self.hparams.learning_rate,
-            betas=self.hparams.betas,
-            eps=self.hparams.eps,
-            weight_decay=self.hparams.weight_decay,
-        )
-        scheduler = torch.optim.lr_scheduler.ExponentialLR(optimizer, gamma=0.999875)
-        return [optimizer], [scheduler]
+    return merged
 
 
 class ForwardTTSTrainingEngine(BaseTrainingEngine):
@@ -425,16 +121,12 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
         config: TrainingEngineConfig,
         dataset_paths: List[Path],
         **kwargs: Any,
-    ) -> pl.LightningModule:
-        extra = dict(config.extra)
-        extra.setdefault("variant", self.default_variant)
+    ) -> "pl.LightningModule":
+        from phoonnx_train.fastpitch.lightning import ForwardTTSModule
+
         return ForwardTTSModule(
-            num_symbols=config.num_symbols,
-            num_speakers=config.num_speakers,
-            sample_rate=config.sample_rate,
             dataset=[str(p) for p in dataset_paths],
-            **extra,
-            **kwargs,
+            **resolve_module_kwargs(config, self.default_variant, **kwargs),
         )
 
     def export_onnx(
@@ -448,9 +140,16 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
 
         Contract matches ``coqui_fastpitch_export/export_fp.py`` /
         ``phoonnx.engines.fastpitch.FastPitchAdapter``: ``token_ids`` [B,T]
-        (+ optional ``speaker`` [B]) -> ``mel_spec`` [B, 80, T_mel].
+        + ``pace``/``pitch_mul``/``pitch_add`` [1] (+ optional ``speaker``
+        [B]) -> ``mel_spec`` [B, 80, T_mel].
         """
         import json
+
+        import torch
+
+        from phoonnx_train.fastpitch.lightning import ForwardTTSModule
+        from phoonnx_train.fastpitch.model import ForwardTTS
+        from phoonnx_train.torch_compat import onnx_export_kwargs
 
         with open(config_path, "r", encoding="utf-8") as f:
             model_config: Dict[str, Any] = json.load(f)
@@ -479,8 +178,8 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
                        pace: torch.Tensor,
                        pitch_mul: torch.Tensor,
                        pitch_add: torch.Tensor,
-                       speaker: Optional[torch.Tensor] = None) -> torch.Tensor:
-                sid = speaker if self.multispeaker else None
+                       speaker=None) -> torch.Tensor:
+                sid = speaker.long() if self.multispeaker else None
                 return self.m.inference(token_ids, speaker=sid, pace=pace,
                                         pitch_mul=pitch_mul, pitch_add=pitch_add)
 
@@ -491,8 +190,10 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
         dummy_controls = (torch.ones(1), torch.ones(1), torch.zeros(1))
         control_names = ["pace", "pitch_mul", "pitch_add"]
         if multispeaker:
-            dummy_speaker = torch.zeros(1, dtype=torch.long)
-            dummy_args: Tuple[Any, ...] = (dummy_ids, *dummy_controls, dummy_speaker)
+            # int32 to match the FastPitchAdapter/MixerTTSAdapter feed dtype;
+            # cast to long inside the wrapper for the embedding lookup
+            dummy_speaker = torch.zeros(1, dtype=torch.int32)
+            dummy_args = (dummy_ids, *dummy_controls, dummy_speaker)
             input_names = ["token_ids", *control_names, "speaker"]
             dynamic_axes = {
                 "token_ids": {0: "batch_size", 1: "phonemes"},
@@ -510,21 +211,17 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
         output_dir.mkdir(parents=True, exist_ok=True)
         output_path = output_dir / f"{checkpoint_path.stem}.onnx"
 
-        export_kwargs: Dict[str, Any] = dict(
-            input_names=input_names,
-            output_names=["mel_spec"],
-            dynamic_axes=dynamic_axes,
-            opset_version=OPSET_VERSION,
-        )
-        # torch>=2.5 defaults torch.onnx.export to the dynamo-based exporter,
-        # which additionally requires the optional 'onnxscript' package.
-        # Force the legacy TorchScript-tracing exporter (matches
-        # coqui_fastpitch_export/export_fp.py) when the kwarg is available.
-        if "dynamo" in torch.onnx.export.__code__.co_varnames:
-            export_kwargs["dynamo"] = False
-
         with torch.no_grad():
-            torch.onnx.export(wrapper, dummy_args, str(output_path), **export_kwargs)
+            torch.onnx.export(
+                wrapper,
+                dummy_args,
+                str(output_path),
+                input_names=input_names,
+                output_names=["mel_spec"],
+                dynamic_axes=dynamic_axes,
+                opset_version=OPSET_VERSION,
+                **onnx_export_kwargs(),
+            )
 
         try:
             import onnx as _onnx
@@ -532,7 +229,7 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
             onnx_model = _onnx.load(str(output_path))
             del onnx_model.metadata_props[:]
             for key, value in {
-                "model_type": "fastpitch" if module.hparams.variant == "fastpitch" else "speedyspeech",
+                "model_type": module.hparams.variant,
                 "engine": module.hparams.variant,
                 "n_speakers": num_speakers,
                 "n_vocab": num_symbols,
@@ -601,12 +298,14 @@ class ForwardTTSTrainingEngine(BaseTrainingEngine):
 
     def load_checkpoint(
         self,
-        model: pl.LightningModule,
+        model: "pl.LightningModule",
         checkpoint_path: Path,
         **kwargs: Any,
-    ) -> pl.LightningModule:
+    ) -> "pl.LightningModule":
         """Tolerant checkpoint load (skips shape-mismatched tensors, e.g.
         when fine-tuning onto a different phoneme inventory)."""
+        import torch
+
         ckpt = torch.load(checkpoint_path, map_location="cpu", weights_only=True)
         state_dict = ckpt.get("state_dict", ckpt)
         model_state = model.state_dict()
